Switch how services are matched
Previously, we looped through each API location, then through every local_service to find matches.

Instead, we clean up the API locations to remove those not registered. Then we loop through every local_service to find matches in the API locations and remove them.

We end up with just the API locations that are not in the dbase.